+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 ###Updated
 - Fix minimum requirement of Symfony at Symfony 3.0
 
+###Added
+- LoggerInterface support in RequestBridge to log request result and errors during execution
+- Default logger to write request in stdout and error in stderr
+
 ##[0.0.1-alpha3] - 2017-03-29
 ###Fixed
 - Enable deep cloning into RequestBridge class to avoid Symfony's Kernel sharing betwin requests. Fix #1